\name{BRCmap-package}
\alias{BRCmap-package}
\alias{BRCmap}
\docType{package}
\title{
BRCmap
}
\description{
An R package to facilitate species distribution mapping as used in Species Distribution Atlases. At present the package is aimed at British Recording however in theory the package can be expanded to cover other regions, with the addition of functions to deal with additional coordinate system, shape files.
}
\details{
\tabular{ll}{
Package: \tab BRCmap\cr
Type: \tab Package\cr
Version: \tab 0.10.2\cr
Date: \tab 2014-06-02\cr
License: \tab GPL-3\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Colin Harrower

Maintainer: Colin Harrower <corr@ceh.ac.uk>

}
\references{

}
\keyword{ package }

\seealso{

}
\examples{
## Examples to come later
}