% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRCmap.r
\docType{package}
\name{BRCmap}
\alias{BRCmap}
\title{BRCmap: A package for producing atlas style maps for UK & Irish 
Biological Recording data with locations as grid reference strings}
\description{
The BRCmap package provides a number of different functions to deal
with and map UK, Irish and Channel Islands grid reference strings.
}
