% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUK_gr_symbols.r
\name{draw_symbols}
\alias{draw_symbols}
\title{Add symbols to existing plot}
\usage{
draw_symbols(x, y, symbol = "circle", dimen, rot = 0, n_circle = 32, ...)
}
\arguments{
\item{x}{a numerical vector specifying the x coordinates of the central
points for each symbol.}

\item{y}{a numerical vector specifying the y coordinates of the central
points for each symbol.}

\item{symbol}{A text string specifying the type of symbols to be draw. The 
default \code{symbol = "circle"} draws a circle with radius dimen/2. Other
possible values are \code{triangle}, \code{square}, \code{diamond}.}

\item{dimen}{The dimensions of the symbol. This is given as the maximum 
height/length of the total polygon and not the radii from the central point}

\item{rot}{A numerical value specifying the rotation of the symbol in
degrees. The default \code{rot = 0} means that the symbol will be plotted
from 12 o'clock working around the central point going clockwise.  An 
exception to this is when \code{symbol = "square"} in which case the
vertices are rotated by 45 degrees so that the top edge is parallel with 
the x axis.}

\item{n_circle}{A number specifying the number number of vertices in the
polygon that will be used to approximate a polygon. The default
\code{n_circle = 32} will give a good approximation of a circle in most
cases, however if you require smooth approximations then inccrease this
value.}

\item{...}{any other arugments will be passed through the \code{polygon} 
function used to actually add the polygon to the plot, e.g. \code{col = 
"red"}, code{border = "blue"}, etc.}
}
\description{
\code{draw_symbols} draws symbols of specified type and size
  onto the existing plot
}
\details{
This function is used by functions in BRCmap (see See Also seciton
  below) to place plotting symbols of a given size at coordinates specified.
  At present only a couple of symbols (circle, triangle, square, diamond) are
  possible at present. Other commonly used mapping symbols will be added in
  the future.
  
  It should be noted that the 'circles' produced are actually n-vertice
  polygons that in most cases should give acceptable approximations for a
  true circle. If you require a smoother circle increase the number of
  vertices in the polygon used to approximate the circle (via
  \code{n_circle}). It should be noted that symbols can appear more jagged in
  the default R device that they will when plotted to other devices such as
  pdf.
}
\examples{
  ## Example of adding symbols to a plot
  # First create a blank plot
    plot(1:60,1:60, type="n")
  # Now add a series of pentagons across the diagonal
    xy_vals = seq(10,50,by=10)
    draw_symbols(xy_vals, xy_vals, "circle", dimen = 3, col="red")
    
}
\seealso{
\code{\link{plotUK_gr_symbols}}
}
\author{
Colin Harrower
}
