% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTMtoMGRS.R
\name{gps_latlon2mgrs}
\alias{gps_latlon2mgrs}
\title{Convert WGS84 Latitude & Longitude values to Miltary Grid Reference System (MGRS)}
\usage{
gps_latlon2mgrs(lat, lon, output_type = "mgrs")
}
\arguments{
\item{lat}{a numerical vector containing latitude values}

\item{lon}{a numerical vector containing longitude values}
}
\value{
A character vector containing the MGRS grid references
}
\description{
This function takes vectors of latitude & longitude values and converts them 
to Miltary Grid Reference System (MGRS) grid references
}
\details{
This function is essentially a wrapper function that uses other functions 
within \code{BRCmap} to do the conversion (listed in See Also section).The vectors supplied to \code{lat} and 
\code{lon} need to be paired, so that 1st values from each are the latitude and 
longitude coordinates for the 1st WGS84 point, the 2nd values from each vector are
the coordiantes for the 2nd WGS84 point, and so on.
}
\examples{
## Converting WGS84 latitude and longitude values to MGRS grid references
   gps_latlon2mgrs(51.602382,-1.110548)
   
}
\seealso{
\code{\link{UTMtoMGRS}}, \code{\link{gps_latlon2UTM}}
}
\author{
Colin Harrower
}
