% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_latlon2UTM.r
\name{gps_latlon2utm}
\alias{gps_latlon2utm}
\title{Convert WGS84 Latitude & Longitude to UTM grid reference}
\usage{
gps_latlon2utm(lat, lon, to_metre = TRUE, out_string = TRUE)
}
\arguments{
\item{lat}{Numerical value giving the Latitude (in decimal format)}

\item{lon}{Numerical value giving the Longitude (in decimal format)}

\item{to_metre}{Should the UTM grid reference be rounded down to metre
precision. The default \code{to_metre = TRUE} will ensure that the UTM grid
reference(s) are rounded down to metre precision.}

\item{out_string}{A logical value determining whether the results should only
be the UTM grid reference character string(s) or alternatively a data.frame
with columns for each individual components that make up the UTM string
(e.g. zone, hemispwhere, easting, northing). The default \code{out_string =
TRUE} will return onyl the UTM grid references strings, while
\code{out_string = FALSE} will return the data.frame.}
}
\value{
A vector of strings or a data.frame depending on the values of `out_string`
}
\description{
Convert WGS84 Latitude & Longitude to UTM grid reference
}
\examples{



}
