% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr2sf_poly_points.R
\name{gr2sf_poly}
\alias{gr2sf_poly}
\title{Convert grid references to spatial simple features polygon object}
\usage{
gr2sf_poly(gridref, gr_atts = NULL, sep_proj_layers = FALSE, crs = NULL)
}
\arguments{
\item{gridref}{A character vector containing the grid references to be
converted.}

\item{gr_atts}{A data.frame of associated attribute information or data to be
attached to the simple features polygons. The rows of data in the
data.frame must be correspond to the grid references given in
\code{gridref}, e.g. the 1st row in \code{gr_atts} are the data for the 1st
grid ref in \code{gridref}, the 2nd row the data for the 2nd grid ref and
so on. The default \code{NULL} will result in a single attribute GRIDREF,
containing the grid reference string, being associated with each polygon.}

\item{sep_proj_layers}{A logical variable determining what happens when
\code{gridref} contains grid references from different coordinate reference
systems (CRS) and no output CRS is specified (when \code{crs = NULL}). see
\code{details} for more information.}

\item{crs}{A valid coordinate reference system reference that will be
correctly interpreted by \code{\link[sf]{st_crs}}.}
}
\value{
This function will return either a single simple features polygon
  object or a list of simple features polygon objects. The output will depend
  on the grid reference type(s) of the grid references supplied to
  \code{gridref}, whether a Coordinate Reference system is supplied to
  \code{crs}, and or whether \code{sep_proj_layers} is \code{TRUE} or
  code{FALSE}
}
\description{
Takes one or more grid references, and any associated data or
  attributes, and converts them to a simple features polygon object.
}
\details{
Each grid reference in \code{gridref} will be converted to a polygon
  in the simple features object. This means that if \code{gridref} contains
  duplicated grid references multiple overlapping polygons will be present in
  the final simple features object. This may be desirable if the attribute
  information attached to each instance of the same grid references is
  different, but may also be undesirable in some situations. If you want a
  single polygon for each grid reference then you will need to process your
  grid reference and attribute data first to ensure there is only a single
  entry for each gird reference.

  When \code{sep_project_layers = FALSE} a single simple feature polygon
  object will be returned using either the CRS for the grid reference system
  most frequently represented in \code{gridref} or the CRS specified by
  \code{crs} when \code{crs} is not \code{NULL}. When
  \code{sep_project_layers = TRUE} a list of simple features will be returned
  with each element containing a simple feature polygon containing all grid
  references in \code{gridref} coming a single grid ref system (e.g. OSGB,
  OSI, UTM30), and using the corresponding crs, with the number of elements
  in the list depending on the number of grid references types present in
  \code{gridref}.
  
  These functions to convert to simple feature (sf) objects will ultimately
  replace the old legacy functions that convert grid reference data to 
  spatial objects using the \code{sp} package.
}
\examples{

## Example converting a a series of grid references to a simple features polygon object
 # Create a vector of grid references
   gridrefs <- c("SP49","WV57","H40","NN29","NT20","TL97")

 # Create a data.frame of associated data
   gr_attributes = data.frame(site = paste("site",1:length(gridrefs)), VC = c(55,113,230,97,72,26))

 # convert grid refs to simple features object
   gr_sf = gr2sf_poly(gridrefs, gr_atts = gr_attributes)
   
}
\seealso{
\code{\link{gr2sf_points}}
}
\author{
Colin Harrower
}
