% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgrs2utm.r
\name{mgrs2utm}
\alias{mgrs2utm}
\title{Convert MGRS grid reference to UTM grid reference}
\usage{
mgrs2utm(mgrs, centre = FALSE, return_precision = FALSE)
}
\arguments{
\item{mgrs}{Character vector containing MGRS grid reference(s) to be converted to UTM grid references}

\item{centre}{Logical value determining whether the UTM grid reference(s) should be estimated using the centre point of the MGRS grid reference square or the lower/left corner}

\item{return_precision}{Logical value determining}
}
\value{
A vector of UTM grid reference(s) or if \code{return_precision = TRUE} a data.frame containing the UTM grid refences in addition to the precision (in metres) of the grid references.
}
\description{
A function to take one or more grid references using the Military Grid Reference System (MGRS) and to convert them to grid references using the Universal Transverse Mercator (UTM) system.
}
\examples{
## Using the mgrs2utm function

mgrs2utm("30UWV6548")

## Outside of the UK
## Avery Island, Louisiana (home of Tabasco Sauce)
## MGRS gridref 15RXP0509
mgrs2utm("15RXP0509")

}
\seealso{
\code{\link{MGRStoUTM}}
}
