% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UK_countries}
\alias{UK_countries}
\title{A dataset containing detailed boundaries of the countries in the UK,
  Ireland and the Channel Islands for use as a background to species maps}
\format{
A \code{Simple Features data.frame} object containing of the polygons
  required to plot the outlines of the countires comprising the UK, Ireland
  and Channel Islands. The attribute data table associated with the polygons
  has the following fields which can be used to plot subsets of the whole.
  \describe{ \item{COUNTRY}{A field contain a factor stating the country to
  which each feature is attributed. Values in this field are: England,
  Scotland, Wales, Northern Ireland, Republic of Ireland, Bailiwick of
  Guerynsey, Bailiwick of Jersey, Isle of Man} \item{REGION}{A field
  containing a factor stating the region to which each feature is attributed
  with four possible values; Great Britain, Ireland, Isle of Man, and Channel
  Islands} \item{GOVERNMENT}{a factor containing determining the Government
  region the features belong to (United Kingom, Republic of Ireland, Crown
  Dependencies)}\item{ID}{A field contain numerical IDs} }
}
\source{
\url{https://nbn.org.uk/tools-and-resources/nbn-toolbox/watsonian-vice-county-boundaries/}
for Great Britain Watsonian Vice County Boundaries.\cr
\url{https://github.com/SK53/Irish-Vice-Counties} for Irish Watsonian Vice
County Boundaries.\cr
\url{https://www.openstreetmap.org/export#map=10/49.3980/-2.2179} for Channel
Islands boundaries.
}
\usage{
UK_countries
}
\description{
A dataset containing detailed high resolution boundaries for the
  countries within the UK, Ireland and the Channel Islands as a series of
  multipolygon features to be used as the base of maps produced by
  \code{BRCmap}. This spatial object is a detailed high resolution outline
  and may contain more detail than required when mapping large areas and or
  producing maps at limited sizes and depending on the plotting device being
  used can result in notably larger output files, in which case it may be
  better to use the lower resolution version \code{UK_countries_lowres}. The
  coodrinates system used in the dataset is the EPSG:27700 - OSGB 1936 /
  British National Grid - Projected Coodinate References System (CRS),
  therefore the Irish data and Channel Islands data have been reprojected.
  The data in this dataset, along with all other coastline datasets included
  with BRCmap (see links below), are derived from the \code{UK_VC} dataset
  where different datasets were derived by merging and/or simplifying the
  features to create the each dataset. This dataset is version of the
  \code{UK_VC} datasets where the features have been simplified using GIS
  tools before merging the features till there is a single feature for each
  country (see format section below) in the dataset. The \code{UK_VC} dataset
  used to produce the other datasets included with this package was created
  using a couple of different open source datasets (see source section below).
}
\seealso{
\code{\link{UK_countries_lowres}}, \code{\link{UK_VC}}, \code{\link{UK}}
}
\author{
Colin Harrower
}
\keyword{datasets}
