\name{BRCmap-package}
\alias{BRCmap-package}
\alias{BRCmap}
\docType{package}
\title{
BRC Mapping Package
}
\description{
This package provides several tools to help with the mapping of species distributions, particularly with the production of species maps of the sort that are used in distribution Atlases.
}
\details{
\tabular{ll}{
Package: \tab BRCmap\cr
Type: \tab Package\cr
Version: \tab 0.0.4\cr
Date: \tab 2014-02-12\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Colin Harrower

Maintainer: Colin Harrower <corr@ceh.ac.uk>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}