\name{BRCmap-package}
\alias{BRCmap-package}
\alias{BRCmap}
\docType{package}
\title{
BRCmap
}
\description{
An R package to facilitate species distribution mapping as used in Species Distribution Atlases. At present the package is aimed at British Recording however in theory the package can be expanded to cover other regions, with the addition of functions to deal with additional coordinate system, and shape files for those regions.
}
\details{
\tabular{ll}{
Package: \tab BRCmap\cr
Type: \tab Package\cr
Version: \tab 0.10.3.1\cr
Date: \tab 2016-05-10\cr
License: \tab GPL-3\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Colin Harrower

Maintainer: Colin Harrower <corr@ceh.ac.uk>

}
\references{

}
\keyword{ package }

\seealso{

}
\examples{
## Examples to come later
}