\name{BRCmap-package}
\alias{BRCmap-package}
\alias{BRCmap}
\docType{package}
\title{
BRCmap
}
\description{
An R package to facilitate species distribution mapping as used in Species Atlases. At present the package is aimed at British Recording schemes however in theory the package can be expanded to cover other regions.
}
\details{
\tabular{ll}{
Package: \tab BRCmap\cr
Type: \tab Package\cr
Version: \tab 0.10.1\cr
Date: \tab 2014-12-15\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Colin Harrower

Maintainer: Colin Harrower <corr@ceh.ac.uk>

}
\references{

}
\keyword{ package }

\seealso{

}
\examples{
## Examples to come later
}